# spec/openapi.yaml
openapi: 3.1.0
info:
  title: InspoBox API
  description: |
    灵感收集器 - 个人小说灵感收集与管理 API
    
    ## 功能特性
    - ✨ 灵感捕捉 - 快速记录创作灵感
    - 📚 故事管理 - 按故事组织灵感碎片
    - 🏷️ 标签系统 - 多维度分类
    - 🤖 AI 辅助 - 智能重组与创作建议
    - 👤 多用户支持 - 数据完全隔离
    
  version: 1.0.0
  contact:
    name: InspoBox Team
servers:
  - url: http://localhost:3000/api
    description: 本地开发环境
  - url: https://inspobox.vercel.app/api
    description: 生产环境
security:
  - BearerAuth: []
  - CookieAuth: []
...
paths:
...
  /fragments/batch-update:
    post:
      tags:
        - Fragments
      summary: 批量更新碎片
      description: 批量移动碎片到指定故事或添加/移除标签
      operationId: batchUpdateFragments
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchUpdateFragmentsRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  updated:
                    type: integer
                    description: 更新的碎片数量
                  failed:
                    type: integer
                    description: 失败的碎片数量
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /search:
    get:
      tags:
        - Search
      summary: 搜索灵感碎片
      description: |
        全文搜索灵感碎片，支持按内容、故事名、标签名搜索。
        结果按相关性和时间排序。
      operationId: searchFragments
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 200
          description: 搜索关键词
        - name: story_id
          in: query
          schema:
            type: string
          description: 限制搜索范围到指定故事
        - name: tag_ids
          in: query
          schema:
            type: array
            items:
              type: string
          description: 限制搜索范围到指定标签
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
          description: 返回结果数量
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
          description: 分页偏移量
      responses:
        '200':
          description: 搜索结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /search/suggestions:
    get:
      tags:
        - Search
      summary: 获取搜索建议
      description: 根据输入内容返回搜索建议（标签、故事）
      operationId: getSearchSuggestions
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 50
          description: 搜索前缀
      responses:
        '200':
          description: 搜索建议
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchSuggestions'

  # ========== AI History ==========
...
  /relations/suggestions:
    get:
      tags:
        - Relations
      summary: 获取关联推荐
      description: 获取 AI 推荐的碎片关联建议
      operationId: getRelationSuggestions
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
            maximum: 50
      responses:
        '200':
          description: 关联推荐列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RelationSuggestions'
  /relations:
    get:
      tags:
        - Relations
      summary: 获取所有关联
      description: 获取用户确认的所有碎片关联
      operationId: getRelations
      responses:
        '200':
          description: 关联列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FragmentRelation'

    post:
      tags:
        - Relations
      summary: 创建关联
      description: 手动创建碎片之间的关联
      operationId: createRelation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRelationRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FragmentRelation'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: Supabase JWT Token
    CookieAuth:
      type: apiKey
      in: cookie
      name: sb-access-token
  schemas:
...
    Fragment:
      type: object
      required:
        - id
        - content
        - user_id
        - sort_order
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
          description: 碎片唯一标识
        content:
          type: string
          minLength: 1
          maxLength: 10000
          description: 灵感内容
        story_id:
          type: string
          format: uuid
          nullable: true
          description: 所属故事 ID
        story:
          $ref: '#/components/schemas/Story'
          description: 关联的故事信息
        tags:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
          description: 关联的标签列表
        user_id:
          type: string
          format: uuid
          description: 创建者 ID
        sort_order:
          type: integer
          default: 0
          description: 排序权重
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Story:
      type: object
      required:
        - id
        - title
        - color
        - user_id
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
          maxLength: 500
          nullable: true
        color:
          type: string
          pattern: '^#[0-9a-fA-F]{6}$'
          description: 故事主题色（十六进制）
          example: "#3b82f6"
        user_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Tag:
      type: object
      required:
        - id
        - name
        - color
        - user_id
        - created_at
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          minLength: 1
          maxLength: 50
          description: 标签名称（唯一）
        color:
          type: string
          pattern: '^#[0-9a-fA-F]{6}$'
          description: 标签颜色
          example: "#10b981"
        user_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time

    # ========== AI Related ==========
...
    SearchResult:
      type: object
      required:
        - fragments
        - total
        - query
      properties:
        fragments:
          type: array
          items:
            $ref: '#/components/schemas/Fragment'
        total:
          type: integer
          description: 总结果数
        query:
          type: string
          description: 搜索关键词
        has_more:
          type: boolean
          description: 是否有更多结果
    SearchSuggestions:
      type: object
      required:
        - stories
        - tags
      properties:
        stories:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                format: uuid
              title:
                type: string
              color:
                type: string
        tags:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                format: uuid
              name:
                type: string
              color:
                type: string

    # ========== Request Bodies ==========
...
    BatchUpdateFragmentsRequest:
      type: object
      required:
        - ids
        - action
      properties:
        ids:
          type: array
          items:
            type: string
            format: uuid
          minItems: 1
          maxItems: 100
          description: 要更新的碎片 ID 列表
        action:
          type: string
          enum: [move_to_story, add_tags, remove_tags]
          description: 批量操作类型
        story_id:
          type: string
          format: uuid
          nullable: true
          description: 目标故事 ID（action 为 move_to_story 时必填）
        tag_ids:
          type: array
          items:
            type: string
            format: uuid
          description: 要添加/移除的标签 ID 列表（action 为 add_tags/remove_tags 时必填）

    # ========== Relation Schemas ==========
    FragmentRelation:
      type: object
      required:
        - id
        - source_fragment_id
        - target_fragment_id
        - relation_type
        - user_id
        - created_at
      properties:
        id:
          type: string
          format: uuid
        source_fragment_id:
          type: string
          format: uuid
          description: 源碎片 ID
        target_fragment_id:
          type: string
          format: uuid
          description: 目标碎片 ID
        relation_type:
          type: string
          enum: [similar, contrast, sequence, causal, thematic, emotional, reference, custom]
          description: 关联类型
        strength:
          type: number
          minimum: 0
          maximum: 1
          description: 关联强度 0-1
        description:
          type: string
          description: 关联描述
        ai_generated:
          type: boolean
          description: 是否由 AI 生成
        ai_confidence:
          type: number
          minimum: 0
          maximum: 1
          description: AI 置信度
        user_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateRelationRequest:
      type: object
      required:
        - source_fragment_id
        - target_fragment_id
        - relation_type
      properties:
        source_fragment_id:
          type: string
          format: uuid
        target_fragment_id:
          type: string
          format: uuid
        relation_type:
          type: string
          enum: [similar, contrast, sequence, causal, thematic, emotional, reference, custom]
        strength:
          type: number
          minimum: 0
          maximum: 1
        description:
          type: string
...
    RelationSuggestion:
      type: object
      required:
        - id
        - source_fragment_id
        - target_fragment_id
        - relation_type
        - confidence
      properties:
        id:
          type: string
        source_fragment_id:
          type: string
          format: uuid
        target_fragment_id:
          type: string
          format: uuid
        source_fragment:
          type: object
          properties:
            id:
              type: string
            content:
              type: string
        target_fragment:
          type: object
          properties:
            id:
              type: string
            content:
              type: string
        relation_type:
          type: string
          enum: [similar, contrast, sequence, causal, thematic, emotional, reference]
        confidence:
          type: number
          minimum: 0
          maximum: 1
        reason:
          type: string
          description: AI 推荐理由
        status:
          type: string
          enum: [pending, accepted, rejected]
    RelationSuggestions:
      type: object
      required:
        - suggestions
        - total_pending
      properties:
        suggestions:
          type: array
          items:
            $ref: '#/components/schemas/RelationSuggestion'
        total_pending:
          type: integer
          description: 待处理的推荐数量
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
        message:
          type: string
        code:
          type: string
  responses:
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Bad Request"
            message: "Content is required"
            code: "VALIDATION_ERROR"
    Unauthorized:
      description: 未授权，需要登录
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Unauthorized"
            message: "Authentication required"
            code: "AUTH_REQUIRED"
...